feat(panel): deprecate `Panel` in favor of `PanelV2`
BREAKING CHANGE: `Panel` has been deprecated in favor of `PanelV2`
