+++ src/components/Panel/_index.scss
@@ -4,9 +4,6 @@
 /// @copyright IBM Security 2019
 ////
 
-@import '../Button/index';
-@import '../IconButton/index';
-
 @import '../Component/mixins';
 
 @import 'mixins';
