+++ src/components/Panel/index.js
@@ -3,7 +3,6 @@
  * @copyright IBM Security 2019
  */
 
-export Panel from './Panel';
-export PanelContainer from './PanelContainer';
+export default from './Panel';
+
 export PanelContent from './PanelContent';
-export PanelController from './PanelController';
