--- src/components/Panel/migrate-to-2.x.md
@@ -1,5 +1,0 @@
-# `Panel`
-
-## `PanelContainer`
-
-Also refer to [`DataDecorator`](../DataDecorator/migrate-to-2.x.md).
