--- src/components/PanelV2/_index.scss
@@ -1,13 +1,0 @@
-////
-/// Panel v2 component.
-/// @group panelv2
-/// @copyright IBM Security 2019
-////
-
-@import '../Component/mixins';
-
-@import 'mixins';
-
-@include security--component($name: panelv2) {
-  @include panelv2;
-}
