--- src/components/PanelV2/_mocks_/index.js
@@ -1,21 +1,0 @@
-/**
- * @file Panel v2 mocks.
- * @copyright IBM Security 2019
- */
-
-import { className, label } from '../../_mocks_';
-
-const active = true;
-
-const labels = {
-  /** @type {string} The panel container's close button `aria-label` */
-  PANEL_CONTAINER_CLOSE_BUTTON: 'Close - labels',
-};
-
-/**
- * Empty click handler.
- * @returns {Object} An empty object.
- */
-const onClick = () => ({});
-
-export { active, className, label, labels, onClick };
