--- src/components/PanelV2/index.js
@@ -1,6 +1,0 @@
-/**
- * @file Panel v2 entry point.
- * @copyright IBM Security 2019
- */
-
-export default from './PanelV2';
