+++ src/components/_index.scss
@@ -20,7 +20,6 @@
 @import 'NonEntitledSection/index';
 @import 'Notification/index';
 @import 'Panel/index';
-@import 'PanelV2/index';
 @import 'Pill/index';
 @import 'ProgressIndicator/index';
 @import 'ScrollGradient/index';
